+++ packages/ai/tsup.config.ts
@@ -12,10 +12,6 @@ export default defineConfig([
   // Internal APIs
   {
     entry: ['internal/index.ts'],
-    // This bundle isn't actually used,
-    // we export the internal bundle with @internal from the root package
-    // and provide different types in package.json for the exports
-    // to save duplicating 40kb for bundle size
     outDir: 'dist/internal',
     format: ['cjs', 'esm'],
     dts: true,